resolve todo: Passing entire team array from model TeamParseable
